+++ source/Core/Castor3D/Overlay/OverlayCategory.cpp
@@ -180,14 +180,14 @@ namespace castor3d
 			{
 				changed = changed || ( ptPos[0] != double( pos.x() ) / totalSize[0] );
 				ptPos[0] = pos.x() / totalSize[0];
-				m_computeSize[0] = uint32_t( renderSize[0] );
+				m_computeSize[0] = renderSize[0];
 			}
 
 			if ( pos.y() )
 			{
 				changed = changed || ( ptPos[1] != double( pos.y() ) / totalSize[1] );
 				ptPos[1] = pos.y() / totalSize[1];
-				m_computeSize[1] = uint32_t( renderSize[1] );
+				m_computeSize[1] = renderSize[1];
 			}
 
 			if ( changed )
@@ -211,14 +211,14 @@ namespace castor3d
 			{
 				changed = changed || ( ptSize[0] != double( size.getWidth() ) / totalSize[0] );
 				ptSize[0] = size.getWidth() / totalSize[0];
-				m_computeSize[0] = uint32_t( renderSize[0] );
+				m_computeSize[0] = renderSize[0];
 			}
 
 			if ( size.getHeight() )
 			{
 				changed = changed || ( ptSize[1] != double( size.getHeight() ) / totalSize[1] );
 				ptSize[1] = size.getHeight() / totalSize[1];
-				m_computeSize[1] = uint32_t( renderSize[1] );
+				m_computeSize[1] = renderSize[1];
 			}
 
 			if ( changed )
